requirements.txt
boto3==1.34.127
botocore==1.34.127
gTTS==2.5.1
moviepy==1.0.3
playwright==1.44.0
praw==7.7.1
prawcore~=2.3.0
requests==2.32.3
rich==13.7.1
toml==0.10.2
translators==5.9.2
pyttsx3==2.90
Pillow==10.3.0
tomlkit==0.12.5
Flask==3.0.3
clean-text==0.6.0
unidecode==1.3.8
spacy==3.7.5
torch==2.3.1
transformers==4.41.2
ffmpeg-python==0.2.0
elevenlabs==1.3.0
yt-dlp==2024.5.27
numpy==1.26.4
python-dotenv==1.0.0
